Razorpay payment link generation is sucessfull (#1)

+++ src/main/java/dev/ashish/EcomPaymentService/strategy/RazorpayPaymentLinkResponse.java
@@ -4,10 +4,10 @@ public class RazorpayPaymentLinkResponse implements PaymentLinkResponse{
     private String paymentLink;
     private String paymentId;
     private String currency;
-    private long amount;
+    private double amount;
     private String description;
 
-    public RazorpayPaymentLinkResponse(String paymentLink, String paymentId, String currency, long amount, String description) {
+    public RazorpayPaymentLinkResponse(String paymentLink, String paymentId, String currency, double amount, String description) {
         this.paymentLink = paymentLink;
         this.paymentId = paymentId;
         this.currency = currency;
@@ -31,7 +31,7 @@ public class RazorpayPaymentLinkResponse implements PaymentLinkResponse{
     }
 
     @Override
-    public Long getAmount() {
+    public double getAmount() {
         return amount;
     }
